requirements.txt
boto3==1.34.52
oracledb==2.0.1
python-dotenv==1.0.1
